Minor UI changes and add null check in Delete

+++ Controllers/ContactsController.cs
@@ -92,6 +92,7 @@ namespace Phonebook.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var contact = await _context.Contacts.FindAsync(id);
+            if (contact == null) return NotFound();
             _context.Contacts.Remove(contact);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
